feat: add quick-filter actions to table viz cell menu in edit mode
The table visualization's cell menu (Chart panel in the explorer and
chart edit page) only offered Copy value / underlying data / drill,
while the results table below had Show only / Exclude. Extract the
quick-filter menu items into a shared QuickFilterMenuItems component
and render them in the table viz menu too, gated on edit mode and the
explorer store being available (the same menu renders in contexts
without one, e.g. AI copilot visualizations).
